BACKEND/requirements.txt
openai
pymupdf
PyPDF2
python-dotenv
tiktoken
flask
flask-cors
numpy
scikit-learn
faiss-cpu
docx